rtl: Synthetize RAM inside FPGA
RAM = RAM block with 8-bit write and 8-read data bus and
16-bit write and 16-bit read address bus.

ram_wrapper = adapts the RAM block to CPU. The CPU has only 1 bus for
reading and writing.

This branch is a working example, but it looks pretty ugly. It will be
used for testing purposes until a better implementation of memory comes.

Signed-off-by: Daniel Matyas <matyasdaniel004@gmail.com>